rag-api/pyproject.toml
[project]
name = "rag-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio>=4.12.0",
    "fastapi[standard]>=0.124.0",
    "httpx>=0.28.1",
    "markdown>=3.10",
    "openai>=2.9.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.38.0",
]
